requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
curl_cffi==0.15.0
frozendict==2.4.7
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multitasking==0.0.12
numpy==2.4.4
pandas==3.0.2
peewee==4.0.4
platformdirs==4.9.6
protobuf==7.34.1
pycparser==3.0
Pygments==2.20.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.33.1
rich==14.3.4
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yfinance==1.2.1
python-dotenv==1.2.2

fastapi
uvicorn
pytest
httpx
google-genai
